Adicionado cadastro de  Usuario, funcionando perfeitamente
