+++ source/GamimgTrack/forms.py
@@ -2,4 +2,8 @@ from django import forms
 
 class registerUserForm(forms.Form):
     nome = forms.CharField(label="Nome", max_length=255)
-    check = forms.BooleanField()
+    email = forms.CharField(label="E-mail", max_length=255)
+    login = forms.CharField(required=False)
+    password = forms.CharField(max_length=255, label="senha")
+    password1 = forms.CharField(max_length=255, label="repetir senha")
+    check = forms.BooleanField(label="Você está de acordo com os nossos termos?")
